Add register method to LoginDao, add getter method for WalletDao, Update ConsoleService to allow printing of menus, add functionality to main CLI class.

+++ capstone/tenmo-client/src/main/java/com/techelevator/tenmo/model/Wallet.java
@@ -1,10 +1,12 @@
 package com.techelevator.tenmo.model;
 
+import java.math.BigDecimal;
+
 public class Wallet {
 
     private int id;
     private int userId;
-    private int balance;
+    private BigDecimal balance;
 
     public int getId() {
         return id;
@@ -22,11 +24,11 @@ public class Wallet {
         this.userId = userId;
     }
 
-    public int getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
